+++ HISTORY.md
@@ -3,6 +3,7 @@
 - Logging: fix logging in parallel runs with new joblib loky backend. Thanks to
   Ben Liesfeld and Roland Ewald.
 - RNAseq: gzip/bgzip FASTQ files in `work/fastq` instead of the original directory.
+- use biobambam2 BAM to FASTQ conversion instead of Picard in all cases.
 
 ## 1.1.1 (6 November 2018)